+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Enable FEP-e232 by default.
 - Don't remove paragraphs when editing bio.
 - Sort bookmarks by creation date.
+- Federate 👎 reaction as `Dislike` activity.
 
 ### Removed